requirements.txt
aiofiles
chardet
python-dotenv
openai
ffmpeg-python
